add a validation

+++ src/main/java/com/example/demo/api/controller/PlayerControllerImpl.java
@@ -3,9 +3,10 @@ package com.example.demo.api.controller;
 import com.example.demo.api.request.CreatePlayerRequest;
 import com.example.demo.api.request.PlayerFilter;
 import com.example.demo.api.request.UpdatePlayerRequest;
+import com.example.demo.api.response.BaseResponse;
 import com.example.demo.api.response.CreatePlayerResponse;
 import com.example.demo.api.response.GetPlayerResponse;
-import com.example.demo.api.response.UpdatePlayerResponse;
+import com.example.demo.api.response.NotFoundResponse;
 import com.example.demo.dto.PlayerDto;
 import com.example.demo.entity.Profession;
 import com.example.demo.entity.Race;
@@ -13,6 +14,7 @@ import com.example.demo.filter.PlayerOrder;
 import com.example.demo.mapper.DtoMapper;
 import com.example.demo.service.PlayerService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -30,20 +32,52 @@ public class PlayerControllerImpl implements PlayerController {
         return DtoMapper.convertToCreateResponse(createdPlayer);
     }
 
-    public GetPlayerResponse getPlayerById(int id) {
+    public ResponseEntity<BaseResponse> getPlayerById(int id) {
         PlayerDto playerDto = playerService.getPlayerById(id);
-        return DtoMapper.convertToGetResponse(playerDto);
+
+        BaseResponse baseResponse;
+        HttpStatus httpStatus;
+        if (playerDto == null) {
+            baseResponse = NotFoundResponse.builder()
+                    .message("Player with id = " + id + " was not found")
+                    .build();
+            httpStatus = HttpStatus.NOT_FOUND;
+        } else {
+            baseResponse = DtoMapper.convertToGetResponse(playerDto);
+            httpStatus = HttpStatus.OK;
+        }
+
+        return new ResponseEntity<>(baseResponse, httpStatus);
     }
 
-    public void deletePlayer(int id) {
-        playerService.deletePlayer(id);
+    public ResponseEntity<Void> deletePlayer(int id) {
+        HttpStatus httpStatus;
+        if (playerService.getPlayerById(id) == null) {
+            httpStatus = HttpStatus.NOT_FOUND;
+        } else {
+            playerService.deletePlayer(id);
+            httpStatus = HttpStatus.OK;
+        }
+        return ResponseEntity.status(httpStatus).build();
     }
 
-    public UpdatePlayerResponse updatePlayer(int id, UpdatePlayerRequest updatePlayerRequest) {
+    public ResponseEntity<BaseResponse> updatePlayer(int id, UpdatePlayerRequest updatePlayerRequest) {
         updatePlayerRequest.setId(id);
-        PlayerDto playerDto = DtoMapper.convertToPlayerDto(updatePlayerRequest);
-        PlayerDto updatedPlayer = playerService.updatePlayer(playerDto);
-        return DtoMapper.convertToUpdateResponse(updatedPlayer);
+        BaseResponse baseResponse;
+        HttpStatus httpStatus;
+        if (playerService.getPlayerById(id) == null) {
+            baseResponse = NotFoundResponse.builder()
+                    .message("Player with id = " + id + " was not found")
+                    .build();
+            httpStatus = HttpStatus.NOT_FOUND;
+        } else {
+            PlayerDto playerDto = DtoMapper.convertToPlayerDto(updatePlayerRequest);
+            PlayerDto updatedPlayer = playerService.updatePlayer(playerDto);
+            baseResponse = DtoMapper.convertToUpdateResponse(updatedPlayer);
+            httpStatus = HttpStatus.OK;
+        }
+
+        return new ResponseEntity<>(baseResponse, httpStatus);
     }
 
     public List<GetPlayerResponse> getFilteredPlayers(String name,
@@ -66,8 +100,8 @@ public class PlayerControllerImpl implements PlayerController {
                 .title(title)
                 .race(race)
                 .profession(profession)
-                .before(new Date(before))
-                .after(new Date(after))
+                .before(before == null ? null : new Date(before))
+                .after(after == null ? null : new Date(after))
                 .order(order)
                 .minExperience(minExperience)
                 .maxExperience(maxExperience)
@@ -99,8 +133,8 @@ public class PlayerControllerImpl implements PlayerController {
                 .title(title)
                 .race(race)
                 .profession(profession)
-                .before(new Date(before))
-                .after(new Date(after))
+                .before(before == null ? null : new Date(before))
+                .after(after == null ? null : new Date(after))
                 .minExperience(minExperience)
                 .maxExperience(maxExperience)
                 .minLevel(minLevel)
